+++ src/rx-database.ts
@@ -14,13 +14,15 @@ import type {
     ServerOptions,
     RxDatabaseCreator,
     RxDumpDatabase,
-    RxDumpDatabaseAny
+    RxDumpDatabaseAny,
+    RxCollectionCreatorBase
 } from './types';
 
 import {
     promiseWait,
     pluginMissing,
-    LOCAL_PREFIX
+    LOCAL_PREFIX,
+    flatClone
 } from './util';
 import {
     newRxError
@@ -179,7 +181,112 @@ export class RxDatabaseBase<
     }
 
     /**
+     * creates multiple RxCollections at once
+     * to be much faster by saving db txs and doing stuff in bulk-operations
+     * This function is not called often, but mostly in the critical path at the initial page load
+     * So it must be as fast as possible
+     */
+    async addCollections(collectionCreators: {
+        // TODO instead of [name: string] only allow keyof Collections
+        [name: string]: RxCollectionCreatorBase
+    }): Promise<{ [key: string]: RxCollection }> {
+        const pouch: PouchDBInstance = this.internalStore as any;
+
+        // get local management docs in bulk request
+        const result = await pouch.allDocs({
+            include_docs: true,
+            keys: Object.keys(collectionCreators).map(name => _collectionNamePrimary(name, collectionCreators[name].schema))
+        });
+        const internalDocByCollectionName: any = {};
+        result.rows.forEach(row => {
+            if (!row.error) {
+                internalDocByCollectionName[row.key] = row.doc;
+            }
+        });
+
+        const schemaHashByName: { [k: string]: string } = {};
+        const collections = await Promise.all(
+            Object.entries(collectionCreators).map(([name, args]) => {
+                const internalDoc = internalDocByCollectionName[_collectionNamePrimary(name, collectionCreators[name].schema)];
+                const useArgs: RxCollectionCreator = flatClone(args) as any;
+                useArgs.name = name;
+                const schema = createRxSchema(args.schema);
+                schemaHashByName[name] = schema.hash;
+                (useArgs as any).schema = schema;
+                (useArgs as any).database = this;
+
+                // TODO check if already exists and schema hash has changed
+
+                // collection already exists
+                if ((this.collections as any)[name]) {
+                    throw newRxError('DB3', {
+                        name
+                    });
+                }
+
+                // collection already exists but has different schema
+                if (internalDoc && internalDoc.schemaHash !== schemaHashByName[name]) {
+                    throw newRxError('DB6', {
+                        name: name,
+                        previousSchemaHash: internalDoc.schemaHash,
+                        schemaHash: schemaHashByName[name]
+                    });
+                }
+
+                // run hooks
+                const hookData: RxCollectionCreator = flatClone(args) as any;
+                (hookData as any).database = this;
+                hookData.name = name;
+                runPluginHooks('preCreateRxCollection', hookData);
+
+                return createRxCollection(useArgs);
+            })
+        );
+
+        const bulkPutDocs: any[] = [];
+        const ret: { [key: string]: RxCollection } = {};
+        collections.forEach(collection => {
+            const name = collection.name;
+            ret[name] = collection;
+            if (
+                collection.schema.crypt &&
+                !this.password
+            ) {
+                throw newRxError('DB7', {
+                    name
+                });
+            }
+
+            // add to bulk-docs list
+            if (!internalDocByCollectionName[name]) {
+                bulkPutDocs.push({
+                    _id: _collectionNamePrimary(name, collectionCreators[name].schema),
+                    schemaHash: schemaHashByName[name],
+                    schema: collection.schema.normalized,
+                    version: collection.schema.version
+                });
+            }
+
+            // set as getter to the database
+            (this.collections as any)[name] = collection;
+            if (!(this as any)[name]) {
+                Object.defineProperty(this, name, {
+                    get: () => (this.collections as any)[name]
+                });
+            }
+        });
+
+        // make a single call to the pouchdb instance
+        await pouch.bulkDocs({
+            docs: bulkPutDocs,
+        });
+
+        return ret;
+    }
+
+    /**
      * create or fetch a collection
+     * @deprecated use addCollections() instead, it is faster and better typed
      */
     collection<
         RxDocumentType = any,
@@ -196,88 +303,12 @@ export class RxDatabaseBase<
             return Promise.resolve(this.collections[args]);
         }
 
-        args = Object.assign({}, args);
-
-        (args as any).database = this;
-
-        runPluginHooks('preCreateRxCollection', args);
-
-        if ((this.collections as any)[args.name]) {
-            throw newRxError('DB3', {
-                name: args.name
-            });
-        }
-
-        const internalPrimary = _collectionNamePrimary(args.name, args.schema);
-
-        const schema = createRxSchema(args.schema);
-        args.schema = schema as any;
-
-        // check schemaHash
-        const schemaHash = schema.hash;
-
-        let colDoc: any;
-        let col: any;
-        return this.lockedRun(
-            () => (this.internalStore as any).get(internalPrimary)
-        )
-            .catch(() => null)
-            .then((collectionDoc: any) => {
-                colDoc = collectionDoc;
-
-                if (collectionDoc && collectionDoc.schemaHash !== schemaHash) {
-                    // collection already exists with different schema, check if it has documents
-                    const pouch = this._spawnPouchDB(args.name, args.schema.version, args.pouchSettings);
-                    return pouch.find({
-                        selector: {},
-                        limit: 1
-                    }).then(oneDoc => {
-                        if (oneDoc.docs.length !== 0) {
-                            // we have one document
-                            throw newRxError('DB6', {
-                                name: args.name,
-                                previousSchemaHash: collectionDoc.schemaHash,
-                                schemaHash
-                            });
-                        }
-                        return collectionDoc;
-                    });
-                } else return collectionDoc;
-            })
-            .then(() => createRxCollection(args as any))
-            .then((collection: RxCollection) => {
-                col = collection;
-                if (
-                    collection.schema.crypt &&
-                    !this.password
-                ) {
-                    throw newRxError('DB7', {
-                        name: args.name
-                    });
-                }
-
-                if (!colDoc) {
-                    return this.lockedRun(
-                        () => (this.internalStore as any).put({
-                            _id: internalPrimary,
-                            schemaHash,
-                            schema: collection.schema.normalized,
-                            version: collection.schema.version
-                        })
-                    ).catch(() => { });
-                }
-            })
-            .then(() => {
-                (this.collections as any)[args.name] = col;
-
-                if (!(this as any)[args.name]) {
-                    Object.defineProperty(this, args.name, {
-                        get: () => (this.collections as any)[args.name]
-                    });
-                }
-
-                return col;
-            });
+        // collection() is deprecated, call new bulk-creation method
+        return this.addCollections({
+            [args.name]: args
+        }).then(colObject => {
+            return colObject[args.name] as any;
+        });
     }
 
     /**